# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Credit Repair Pro API",
    "description": "Complete API for Credit Repair SaaS — client management, credit tracking, dispute generation, payments, compliance (CROA/FCRA/GLBA), and AI-powered analysis.",
    "version": "1.0.0",
    "contact": {
      "name": "TriExpert Services",
      "url": "https://triexpertservice.com"
    }
  },
  "servers": [
    { "url": "https://triexpertservice.com/api", "description": "Production" },
    { "url": "http://localhost:5000/api", "description": "Development" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Local JWT or Auth0 token"
      }
    },
...
  },
  "paths": {
...
    "/disputes/client/{clientId}": {
      "get": {
        "tags": ["Disputes"], "summary": "Get all disputes for client", "operationId": "getClientDisputes",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "clientId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": { "200": { "description": "Disputes list" } }
      }
    },
    "/disputes": {
      "post": {
        "tags": ["Disputes"], "summary": "Create dispute with generated letter", "operationId": "createDispute",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["clientId", "creditItemId", "disputeType", "bureau"], "properties": { "clientId": { "type": "string", "format": "uuid" }, "creditItemId": { "type": "string", "format": "uuid" }, "disputeType": { "type": "string" }, "bureau": { "type": "string" } } } } } },
        "responses": { "201": { "description": "Dispute created" } }
      }
    },
    "/disputes/{id}/status": {
      "put": {
        "tags": ["Disputes"], "summary": "Update dispute status/details", "operationId": "updateDisputeStatus",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "status": { "type": "string" }, "sentDate": { "type": "string", "format": "date" }, "responseDate": { "type": "string", "format": "date" }, "responseText": { "type": "string" }, "trackingNumber": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Status updated" } }
      }
    },
    "/disputes/{id}": {
      "get": {
        "tags": ["Disputes"], "summary": "Get single dispute", "operationId": "getDispute",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "Dispute details" }, "404": { "description": "Not found" } }
      }
    },
...
    "/documents/upload": {
      "post": {
        "tags": ["Documents"], "summary": "Upload document", "operationId": "uploadDocument",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "multipart/form-data": { "schema": { "type": "object", "properties": { "document": { "type": "string", "format": "binary" }, "clientId": { "type": "string" }, "disputeId": { "type": "string" }, "documentCategory": { "type": "string" } } } } } },
        "responses": { "201": { "description": "Document uploaded" } }
      }
    },
    "/documents/client/{clientId}": {
      "get": { "tags": ["Documents"], "summary": "Get client's documents", "operationId": "getClientDocuments", "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "clientId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }], "responses": { "200": { "description": "Document list" } } }
    },
    "/documents/{id}/download": {
      "get": { "tags": ["Documents"], "summary": "Download/view a document", "operationId": "downloadDocument", "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }], "responses": { "200": { "description": "File data" } } }
    },
    "/documents/{id}": {
      "delete": { "tags": ["Documents"], "summary": "Delete document + file", "operationId": "deleteDocument", "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }], "responses": { "200": { "description": "Document deleted" } } }
    },
...
  }
}